Move credentials methods from pkg/internal to pkg/openstack
