+++ pkg/internal/infrastructure/infrastructure.go
@@ -17,13 +17,13 @@ package infrastructure
 import (
 	"context"
 
-	"github.com/gardener/gardener-extension-provider-openstack/pkg/internal"
+	"github.com/gardener/gardener-extension-provider-openstack/pkg/openstack"
 
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 // GetCredentialsFromInfrastructure retrieves the ServiceAccount from the Secret referenced in the given Infrastructure.
-func GetCredentialsFromInfrastructure(ctx context.Context, c client.Client, config *extensionsv1alpha1.Infrastructure) (*internal.Credentials, error) {
-	return internal.GetCredentials(ctx, c, config.Spec.SecretRef)
+func GetCredentialsFromInfrastructure(ctx context.Context, c client.Client, config *extensionsv1alpha1.Infrastructure) (*openstack.Credentials, error) {
+	return openstack.GetCredentials(ctx, c, config.Spec.SecretRef)
 }
