+++ pkg/validator/shoot_handler.go
@@ -19,6 +19,8 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/gardener/gardener-extension-provider-openstack/pkg/openstack"
+
 	"github.com/gardener/gardener/extensions/pkg/util"
 	"github.com/gardener/gardener/pkg/apis/core"
 	"github.com/go-logr/logr"
@@ -27,9 +29,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
-
-	"github.com/gardener/gardener-extension-provider-openstack/pkg/internal"
-	"github.com/gardener/gardener-extension-provider-openstack/pkg/openstack"
 )
 
 // Shoot validates shoots
@@ -52,7 +51,7 @@ func (v *Shoot) Handle(ctx context.Context, req admission.Request) admission.Res
 	}
 
 	// Get credentials
-	credentials, err := internal.GetCredentialsBySecretBinding(ctx, v.client, client.ObjectKey{Namespace: shoot.Namespace, Name: shoot.Spec.SecretBindingName})
+	credentials, err := openstack.GetCredentialsBySecretBinding(ctx, v.client, client.ObjectKey{Namespace: shoot.Namespace, Name: shoot.Spec.SecretBindingName})
 	if err != nil {
 		v.Logger.Error(err, "could not get credentials from SecretBindingName %s", shoot.Spec.SecretBindingName)
 		return admission.Errored(http.StatusBadRequest, err)
